+++ Lilla spelprojektet/Code/GameObject.cpp
@@ -52,7 +52,7 @@ void GameObject::mUpdate(ID3D11DeviceContext *dc, std::vector<std::vector<MESH_P
 
 	MESH_PNC *mesh = reinterpret_cast<MESH_PNC*>(mappedData->pData);
 
-	for(int j = 0; j < data.size(); j++)
+	for(int j = 0; j < (int)data.size(); j++)
 	{
 		for(int i = 0; i < (int)data[j].size(); i++)
 		{
@@ -75,7 +75,7 @@ void GameObject::mUpdate(ID3D11DeviceContext *dc , std::vector<HPBarInfo>& data)
 
 	HPBarInfo *mesh = reinterpret_cast<HPBarInfo*>(mappedData->pData);
 
-	for(int j = 0; j < data.size(); j++)
+	for(int j = 0; j < (int)data.size(); j++)
 	{
 		//mesh[j].translate = 
 	}
